# EDIT - 리뷰 받은 수정 사항 반영 기타 수정
- config.hpp 에서 size_t -> int 로 수정
- attachHeader에서 header를 생성할때 Sender_id와 chain_id 의 순서를 바꿔
넣었던것 수정
- attachHeader에서 상수값  변경에 따라 문제가 생길 수 있는 점 수정

+++ src/config/config.hpp
@@ -39,10 +39,10 @@ constexpr uint8_t G = 'G';
 constexpr uint8_t VERSION = '1';
 constexpr uint8_t NOT_USED = 0x00;
 constexpr uint8_t RESERVED[6] = {0x00};
-constexpr size_t HEADER_LENGTH = 32;
-constexpr size_t SENDER_ID_LENGTH = 8;
-constexpr size_t RESERVED_LENGTH = 6;
-constexpr size_t MSG_LENGTH_SIZE = 4;
+constexpr int HEADER_LENGTH = 32;
+constexpr int SENDER_ID_LENGTH = 8;
+constexpr int RESERVED_LENGTH = 6;
+constexpr int MSG_LENGTH_SIZE = 4;
 
 const std::string PORT_NUM = "50051";
 } // namespace config
